mic labels are properly unhardcoded now + seperate RadioDeviceSystem into regions + why did i make those floats again?

+++ Content.Shared/Radio/Components/RadioMicrophoneComponent.cs
@@ -19,7 +19,7 @@ public sealed partial class RadioMicrophoneComponent : Component
 
     // funky - change default listen range (and network it for ui)
     [DataField, AutoNetworkedField]
-    public float ListenRange = 1;
+    public int ListenRange = 1;
 
     /// <summary>
     /// Funky - The maximum listen range
@@ -29,14 +29,14 @@ public sealed partial class RadioMicrophoneComponent : Component
     /// via UI.)
     /// </summary>
     [DataField]
-    public float? MaxRange;
+    public int? MaxRange;
     /// <summary>
     /// Funky - The minimum listen range
     /// this radio can have, if adjustable
     /// by UI. Defaults to 1.
     /// </summary>
     [DataField]
-    public float? MinRange = 1;
+    public int? MinRange = 1;
 
     [DataField]
     public bool Enabled = false;
